- Fixed CompositeBehaviorDescriptionUTest  - Added comments (TODOs) in places where the code rely on uniqueness of hash values (CBDUTest and LocalSpaceMap2D)  - Replaced remaining ocorrences of hash_map/set by boost::unordered_map/set

+++ opencog/embodiment/Learning/behavior/CompositeBehaviorDescription.h
@@ -131,7 +131,7 @@ public:
      */
     static std::vector<unsigned long> *buildSeparatorsVector(const std::vector<ElementaryBehaviorDescription> &entries);
 
-    int hashCode();
+    size_t hashCode();
 
     std::string toString() const;
